chore: Refactor exportFile WF package
