+++ src/main/java/it/gov/pagopa/pu/workflow/service/exportfile/ExportFileServiceImpl.java
@@ -2,7 +2,7 @@ package it.gov.pagopa.pu.workflow.service.exportfile;
 
 import it.gov.pagopa.pu.processexecutions.dto.generated.ExportFile.ExportFileTypeEnum;
 import it.gov.pagopa.pu.workflow.dto.generated.WorkflowCreatedDTO;
-import it.gov.pagopa.pu.workflow.wf.exportfile.ExportFileWFClient;
+import it.gov.pagopa.pu.workflow.wf.exportfile.export.ExportFileWFClient;
 import it.gov.pagopa.pu.workflow.wf.exportfile.expiration.ExportFileExpirationHandlerWFClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
