+++ src/main/java/it/gov/pagopa/pu/workflow/wf/exportfile/export/activity/ScheduleExportFileExpirationActivity.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.pu.workflow.wf.exportfile.activity;
+package it.gov.pagopa.pu.workflow.wf.exportfile.export.activity;
 
 import io.temporal.activity.ActivityInterface;
 import io.temporal.activity.ActivityMethod;
